CrowdFundingCell 컴포넌트화
